Правка 1 задание 4.1 от 17.03.2021
