+++ index.html
@@ -229,7 +229,6 @@
   </template>
 
   <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-  <script src="js/main.js"></script>
-
+  <script src="./js/main.js" defer type="module"></script>
 </body>
 </html>
